Unit Test adjustemnt to follow to new code layout
